Extract duplicated functions for loading tenant to utility module.
