+++ packages/sync-service/lib/electric/plug/delete_shape_plug.ex
@@ -4,7 +4,8 @@ defmodule Electric.Plug.DeleteShapePlug do
 
   alias Electric.Shapes
   alias Electric.Plug.ServeShapePlug.Params
-  alias Electric.TenantManager
+
+  import Electric.Plug.TenantUtils
 
   plug :fetch_query_params
   plug :put_resp_content_type, "application/json"
@@ -24,59 +25,6 @@ defmodule Electric.Plug.DeleteShapePlug do
       |> send_resp(404, Jason.encode_to_iodata!(%{status: "Not found"}))
       |> halt()
     end
-  end
-
-  defp validate_tenant_id(%Plug.Conn{} = conn, _) do
-    case Map.get(conn.query_params, "database_id", :not_found) do
-      :not_found ->
-        conn
-
-      id when is_binary(id) ->
-        assign(conn, :database_id, id)
-
-      _ ->
-        conn
-        |> send_resp(400, Jason.encode_to_iodata!("database_id should be a string"))
-        |> halt()
-    end
-  end
-
-  defp load_tenant(%Plug.Conn{assigns: %{database_id: tenant_id}} = conn, _) do
-    {:ok, tenant_config} = TenantManager.get_tenant(tenant_id, conn.assigns.config)
-    assign_tenant(conn, tenant_config)
-  end
-
-  defp load_tenant(%Plug.Conn{} = conn, _) do
-    # Tenant ID is not specified
-    # ask the tenant manager for the only tenant
-    # if there's more than one tenant we reply with an error
-    case TenantManager.get_only_tenant(conn.assigns.config) do
-      {:ok, tenant_config} ->
-        assign_tenant(conn, tenant_config)
-
-      {:error, :not_found} ->
-        conn
-        |> send_resp(400, Jason.encode_to_iodata!("No database found"))
-        |> halt()
-
-      {:error, :several_tenants} ->
-        conn
-        |> send_resp(
-          400,
-          Jason.encode_to_iodata!(
-            "Database ID was not provided and there are multiple databases. Please specify a database ID using the `database_id` query parameter."
-          )
-        )
-        |> halt()
-    end
-  end
-
-  defp assign_tenant(%Plug.Conn{} = conn, tenant_config) do
-    id = tenant_config[:tenant_id]
-
-    conn
-    |> assign(:config, tenant_config)
-    |> assign(:tenant_id, id)
   end
 
   defp validate_query_params(%Plug.Conn{} = conn, _) do
